refactor : 3차 세미나 과제_service/dto 패키지 구조 엔티티별로 나눔  #7
