+++ week02/seminar/demo/src/main/java/com/example/demo/controller/BlogController.java
@@ -4,8 +4,8 @@ package com.example.demo.controller;
 import com.example.demo.common.dto.SuccessMessage;
 import com.example.demo.common.dto.SuccessStatusResponse;
 import com.example.demo.service.BlogService;
-import com.example.demo.service.dto.BlogCreateRequest;
-import com.example.demo.service.dto.BlogTitleUpdateRequest;
+import com.example.demo.service.dto.blog.BlogCreateRequest;
+import com.example.demo.service.dto.blog.BlogTitleUpdateRequest;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
