+++ week02/seminar/demo/src/main/java/com/example/demo/service/BlogService.java
@@ -5,8 +5,8 @@ import com.example.demo.domain.Blog;
 import com.example.demo.domain.Member;
 import com.example.demo.exception.NotFoundException;
 import com.example.demo.repository.BlogRepository;
-import com.example.demo.service.dto.BlogCreateRequest;
-import com.example.demo.service.dto.BlogTitleUpdateRequest;
+import com.example.demo.service.dto.blog.BlogCreateRequest;
+import com.example.demo.service.dto.blog.BlogTitleUpdateRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
