+++ week02/seminar/demo/src/main/java/com/example/demo/service/MemberService.java
@@ -5,8 +5,8 @@ import com.example.demo.common.dto.ErrorMessage;
 import com.example.demo.domain.Member;
 import com.example.demo.exception.NotFoundException;
 import com.example.demo.repository.MemberRepository;
-import com.example.demo.service.dto.MemberCreateDto;
-import com.example.demo.service.dto.MemberFindDto;
+import com.example.demo.service.dto.member.MemberCreateDto;
+import com.example.demo.service.dto.member.MemberFindDto;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
