+++ week02/seminar/demo/src/main/java/com/example/demo/service/dto/member/MemberFindDto.java
@@ -1,4 +1,4 @@
-package com.example.demo.service.dto;
+package com.example.demo.service.dto.member;
 
 
 import com.example.demo.domain.Member;
